Crate dsa

A tool aiming to provide a cli client for DSA (Das Schwarze Auge).

It provides the ability to parse the Heldensoftware exports and a subset of character related mechanics like:

rolling the dice for you

dumping your character

keeping track of your health, astral points, and stamina

Modules

commands
The commands which can be plugged into the command line App.

error
Error chain created using error-chain.

output
Output formatting.

Structs

Hero
Public container for the hero attributes.

Enums

Quality
Base values of a Hero.

Traits

Action
A command line action providing usage and call functionality.

Functions

app
Helper to create the basic app with options common between the cli-subcommand and the main app.