Trait dsa::output::Formatter

pub trait Formatter {
    fn format(&self, data: &Output) -> String;
}

Converts an Output to a String based on its own rules.

This trait must be implemented by all Formats. Each Format may choose how to represent each Output.

Required methods

fn format(&self, data: &Output) -> String

Convert the Output to a String for further presentation to the user.

Loading content...