Trait dsa::Action

pub trait Action {
    fn usage<'a, 'b>(&'a self) -> App<'b, 'b>;

    fn call(&mut self, hero: &Hero, matches: &ArgMatches) -> Result<Vec<Output>>;
}

A command line action providing usage and call functionality.

The usage() method returns the clap definition of a command line subcommand. The call function is then to be called with the corresponding ArgMatches and the Hero to act on, returning a Result of an arbitrary list of Outputs.

Required methods

fn usage<'a, 'b>(&'a self) -> App<'b, 'b>

Command line argument definition of the subcommand of the action.

This App determines determines the ArgMatches of the call() method.

fn call(&mut self, hero: &Hero, matches: &ArgMatches) -> Result<Vec<Output>>

A method which maps self, as well as the current Hero and the result of the usage()-invocation into zero or more Outputs.

Loading content...